+++ scripts/generate-bmp-diffs.sh
