+++ scripts/pdf-to-bmp.sh
